refactor front end to isolate launchmon dependencies
