+++ src/Makefile.am
@@ -12,11 +12,11 @@ SWIGFLAGS += -DSTAT_GDB_BE
 endif
 AM_LDFLAGS = $(CODE_COVERAGE_LIBS)
 
-include_HEADERS = STAT.h STAT_FrontEnd.h STAT_BackEnd.h STAT_timer.h STAT_GraphRoutines.h STAT_IncMRNet.h MRNetSymbolReader.h STAT_CircularLogs.h
+include_HEADERS = STAT.h STAT_FrontEnd.h STAT_lmonFrontEnd.h STAT_BackEnd.h STAT_timer.h STAT_GraphRoutines.h STAT_IncMRNet.h MRNetSymbolReader.h STAT_CircularLogs.h
 
 if ENABLE_STATBENCH
 stat_bin_programs_statbench = STATBenchbin STATBenchD
-STATBenchbin_SOURCES = STATBench.C STAT.h STAT_FrontEnd.h STAT_GraphRoutines.h
+STATBenchbin_SOURCES = STATBench.C STAT.h STAT_FrontEnd.h STAT_lmonFrontEnd.h STAT_GraphRoutines.h
 STATBenchD_SOURCES = STATBenchD.C STAT.h STAT_BackEnd.h STAT_GraphRoutines.h
 STATBenchbin_LDADD = libstatfe.la @FELIBS@
 STATBenchD_LDADD = libstatbe.la @BELIBS@
@@ -40,13 +40,13 @@ libcallpathwrap_la_CPPFLAGS = -I@ADEPTUTILSPREFIX@/include -I@CALLPATHPREFIX@/in
 libcallpathwrap_la_LDFLAGS = -L@CALLPATHPREFIX@/lib -Wl,-rpath=@CALLPATHPREFIX@/lib -lcallpath -Wl,-init=depositlightcorewrap_init -ldl
 endif
 
-STATbin_SOURCES = STAT.C STAT.h STAT_FrontEnd.h STAT_GraphRoutines.h
+STATbin_SOURCES = STAT.C STAT.h STAT_FrontEnd.h STAT_lmonFrontEnd.h STAT_GraphRoutines.h
 STATD_SOURCES = STATD.C STAT.h STAT_BackEnd.h STAT_GraphRoutines.h
 STATbin_LDADD = libstatfe.la @FELIBS@
 STATD_LDADD = libstatbe.la @BELIBS@
 
 stat_be_la_sources_base = STAT_BackEnd.C STAT_BackEnd.h STAT.h STAT_timer.h STAT_timer.C STAT_CircularLogs.C
-stat_fe_la_sources_base = STAT_FrontEnd.C STAT_FrontEnd.h STAT.h STAT_timer.h STAT_timer.C
+stat_fe_la_sources_base = STAT_FrontEnd.C STAT_lmonFrontEnd.C STAT_FrontEnd.h STAT_lmonFrontEnd.h STAT.h STAT_timer.h STAT_timer.C
 
 libstatbe_la_SOURCES = $(stat_be_la_sources_base) $(stat_graphlib_sources) $(stat_be_la_sources_fgfs)
 libstatfe_la_SOURCES = $(stat_fe_la_sources_base) $(stat_graphlib_sources)
